+++ app/CHANGELOG.md
@@ -153,6 +153,7 @@ In this file you can find what's changed in each version.
 - Include function fixes and TS/docs support in weivData.query function.
 - get function BUG fixes.
 - X.509 method is now allowed and should be working as expected.
+- idConvreter function added to library
 
 ---